Move contacts sample to separate control
Also update the NuGet package to get the latest APIs
