+++ samples/MvvmSample.Core/MvvmSample.Core.csproj
@@ -5,8 +5,7 @@
     <UserSecretsId>4e66f7b4-01a8-4f00-8733-4ae6a08c741f</UserSecretsId>
   </PropertyGroup>
   <ItemGroup>
-    <PackageReference Include="CommunityToolkit.Common" Version="8.0.0-build.63" />
-    <PackageReference Include="CommunityToolkit.Mvvm" Version="8.0.0-build.63" />
+    <PackageReference Include="CommunityToolkit.Mvvm" Version="8.0.0-build.129.g324f7c0f19" />
     <PackageReference Include="Nito.AsyncEx.Coordination" Version="5.1.2" />
     <PackageReference Include="Refit" Version="6.3.2" />
   </ItemGroup>
